+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes for Datastar
 
+## 1.0.0-RC.1 - Unreleased
+
+- The package now requires Datastar [v1.0.0-RC.1](https://github.com/starfederation/datastar/releases/tag/v1.0.0-RC.1).
+- Renamed the `fragments` Blade directive to `patchelements`.
+- Renamed the `removefragments` Blade directive to `removeelements`.
+- Renamed the `defaultFragmentOptions` config setting to `defaultElementOptions`.
+- Removed the `removesignals` Blade directive.
+- Removed the `datastar()->getFragments()` helper method.
+
 ## 1.0.0-beta.8 - 2025-04-15
 
 ### Added